requirements.txt
discord.py
python-dotenv
aiohttp
requests
pydantic
psutil
uvicorn
fastapi